+++ src/Packata.Core/Serialization/Yaml/DataPackageSerializer.cs
@@ -16,13 +16,7 @@ namespace Packata.Core.Serialization.Yaml
         public DataPackage Deserialize(StreamReader reader, HttpClient httpClient, string root)
         {
             var deserializer = new DeserializerBuilder()
-                .WithNamingConvention(CamelCaseNamingConvention.Instance)
-                .WithAttributeOverride(typeof(DataPackage), "Profile", new YamlMemberAttribute { Alias = "$schema" })
-                .WithAttributeOverride(typeof(DataPackage), "Paths", new YamlMemberAttribute { Alias = "path" })
-                .WithAttributeOverride(typeof(Resource), "Profile", new YamlMemberAttribute { Alias = "$schema" })
-                .WithAttributeOverride(typeof(Resource), "Paths", new YamlMemberAttribute { Alias = "path" })
-                .WithAttributeOverride(typeof(Schema), "Profile", new YamlMemberAttribute { Alias = "$schema" })
-                .WithAttributeOverride(typeof(TableDialect), "Profile", new YamlMemberAttribute { Alias = "$schema" })
+                .WithNamingConvention(new DataPackageNamingConvention())
                 .WithTypeConverter(new PathConverter(httpClient, root))
                 .WithTypeConverter(new FieldConverter())
                 .WithTypeConverter(new MissingValuesConverter())
